move tests from bid module to bid_blanks
